+++ src/config/mod.rs
@@ -1,15 +1,16 @@
-use crate::types::Directive;
-use std::str::FromStr;
-
 mod v1;
 mod v2;
 
+/// Configuration as described by the instance of an admonition in markdown.
+///
+/// This structure represents the configuration the user must provide in each
+/// instance.
 #[derive(Debug, PartialEq)]
 pub(crate) struct AdmonitionInfoRaw {
-    directive: String,
-    title: Option<String>,
-    additional_classnames: Vec<String>,
-    collapsible: bool,
+    pub(crate) directive: String,
+    pub(crate) title: Option<String>,
+    pub(crate) additional_classnames: Vec<String>,
+    pub(crate) collapsible: Option<bool>,
 }
 
 /// Extract the remaining info string, if this is an admonition block.
@@ -52,56 +53,6 @@ impl AdmonitionInfoRaw {
     }
 }
 
-#[derive(Debug, PartialEq)]
-pub(crate) struct AdmonitionInfo {
-    pub directive: Directive,
-    pub title: Option<String>,
-    pub additional_classnames: Vec<String>,
-    pub collapsible: bool,
-}
-
-impl AdmonitionInfo {
-    pub fn from_info_string(info_string: &str) -> Option<Result<Self, String>> {
-        AdmonitionInfoRaw::from_info_string(info_string).map(|result| result.map(Into::into))
-    }
-}
-
-impl From<AdmonitionInfoRaw> for AdmonitionInfo {
-    fn from(other: AdmonitionInfoRaw) -> Self {
-        let AdmonitionInfoRaw {
-            directive: raw_directive,
-            title,
-            additional_classnames,
-            collapsible,
-        } = other;
-        let (directive, title) = match (Directive::from_str(&raw_directive), title) {
-            (Ok(directive), None) => (directive, ucfirst(&raw_directive)),
-            (Err(_), None) => (Directive::Note, "Note".to_owned()),
-            (Ok(directive), Some(title)) => (directive, title),
-            (Err(_), Some(title)) => (Directive::Note, title),
-        };
-        // If the user explicitly gave no title, then disable the title bar
-        let title = if title.is_empty() { None } else { Some(title) };
-        Self {
-            directive,
-            title,
-            additional_classnames,
-            collapsible,
-        }
-    }
-}
-
-/// Make the first letter of `input` upppercase.
-///
-/// source: https://stackoverflow.com/a/38406885
-fn ucfirst(input: &str) -> String {
-    let mut chars = input.chars();
-    match chars.next() {
-        None => String::new(),
-        Some(f) => f.to_uppercase().collect::<String>() + chars.as_str(),
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
@@ -121,7 +72,7 @@ mod test {
                 directive: "note".to_owned(),
                 title: None,
                 additional_classnames: vec!["additional-classname".to_owned()],
-                collapsible: false,
+                collapsible: None,
             }
         );
         // v2 syntax is supported
@@ -133,25 +84,7 @@ mod test {
                 directive: "question".to_owned(),
                 title: Some("Custom Title".to_owned()),
                 additional_classnames: Vec::new(),
-                collapsible: false,
-            }
-        );
-    }
-
-    #[test]
-    fn test_admonition_info_from_raw() {
-        assert_eq!(
-            AdmonitionInfo::from(AdmonitionInfoRaw {
-                directive: " ".to_owned(),
-                title: None,
-                additional_classnames: Vec::new(),
-                collapsible: false,
-            }),
-            AdmonitionInfo {
-                directive: Directive::Note,
-                title: Some("Note".to_owned()),
-                additional_classnames: Vec::new(),
-                collapsible: false,
+                collapsible: None,
             }
         );
     }